requirements.txt
Flask==3.1.0
Pillow==12.0.0
gunicorn==21.2.0
opencv-python-headless==4.12.0.88
numpy==2.2.6
scikit-learn==1.7.2
joblib==1.5.2
requests==2.31.0
